refactor: Reorganize simulator commands into its own module
