+++ lib/constants.ts
@@ -1,3 +1,5 @@
+/** Capability value that selects an available simulator automatically. */
+export const UDID_AUTO = 'auto';
 /** Valid values for the `platformName` capability (also used in simulator names). */
 export const PLATFORM_NAME_IOS = 'iOS';
 export const PLATFORM_NAME_TVOS = 'tvOS';
